% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_snps.R
\name{convert_snps}
\alias{convert_snps}
\title{Convert SNPs to ranges}
\usage{
convert_snps(snp_file, output_file, sleep_time)
}
\arguments{
\item{output_file}{Output file of ranges}

\item{snps}{list of SNPs. One per line.}
}
\description{
Convert SNPs to ranges
}
