% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd_codes.R
\name{get_icd_codes}
\alias{get_icd_codes}
\title{Get ICD codes for PheWAS}
\usage{
get_icd_codes(
  table = "workspace_victr.phenotemplate_sd.all_mega_icd_codes_full"
)
}
\arguments{
\item{table}{databricks table that has grid, code, date, vocabulary_id, sex as columns}
}
\description{
Get ICD codes for PheWAS
}
