% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phewas.R
\name{run_phewas}
\alias{run_phewas}
\title{Run PheWAS}
\usage{
run_phewas(
  codes_dat,
  genotypes,
  output_folder,
  n_cores = 8,
  additive.genotypes = FALSE,
  subset1 = NULL
)
}
\arguments{
\item{genotypes}{the file location of the raw plink2_out file produced by run_plink}

\item{output_folder}{output will be stored in <output_folder>/phewas_<Sys.Date()>output_chr_<chromosome>.csv}

\item{n_cores}{number of cores to run on. Default is 8}

\item{additive.genotypes}{should genotypes be additive. Default is FALSE}

\item{subset1}{File containing subset of SNPs from the plink2_out file.}

\item{codes}{dataset with the following columns: id, code, date, vocabulary_id, sex}
}
\description{
Run PheWAS
}
