% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink.R
\name{run_plink}
\alias{run_plink}
\title{Run Plink}
\usage{
run_plink(
  range_file,
  output_folder,
  plink2_input_folder = NULL,
  maf = NULL,
  max_maf_val = NULL
)
}
\arguments{
\item{range_file}{a text file specifying the chromosome, start, end, name values}

\item{output_folder}{plink log and raw will be output to <output_folder>/plink2_out}

\item{plink2_input_folder}{imputed mega data location. Default is "/scratch/mega_data/imputed_mega_data/"}

\item{maf}{filters out all variants with minor allele frequency below the provided threshold (default 0.00005)}

\item{max_maf_val}{imposes an upper MAF bound}
}
\description{
Run Plink
}
