% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink_unimputed.R
\name{run_plink_unimputed}
\alias{run_plink_unimputed}
\title{Run Plink using unimputed data}
\usage{
run_plink_unimputed(
  snp_file,
  output_folder,
  bfile = "/scratch/mega_data/unimputed_mega_data/MEGAv1-1_ALL"
)
}
\arguments{
\item{snp_file}{a text file specifying the SNPs to be extracted}

\item{output_folder}{plink log and raw will be output to <output_folder>/plink2_out}

\item{plink2_input_folder}{imputed mega data location. Default is "/scratch/mega_data/imputed_mega_data/"}

\item{maf}{filters out all variants with minor allele frequency below the provided threshold (default 0.00005)}

\item{max_maf_val}{imposes an upper MAF bound}
}
\description{
Run Plink using unimputed data
}
