% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_to_unimputed_mega.R
\name{snps_to_unimputed_mega}
\alias{snps_to_unimputed_mega}
\title{Convert SNPs to unimputed MEGA call names}
\usage{
snps_to_unimputed_mega(
  snps,
  output_csv,
  output_txt,
  bim_file = "/scratch/mega_data/unimputed_mega_data/MEGAv1-1_ALL.bim",
  sleep_time = 1
)
}
\arguments{
\item{snps}{list of SNPs. One per line.}

\item{output_csv}{Output csv with full info}

\item{output_txt}{Simple text file with new SNPs to be used with Plink}
}
\description{
Convert SNPs to unimputed MEGA call names
}
