% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_snps.R
\name{convert_snps}
\alias{convert_snps}
\title{Convert SNPs to ranges}
\usage{
convert_snps(snps, output_file = NULL, sleep_time = 1)
}
\arguments{
\item{snps}{list of SNPs. One per line.}

\item{output_file}{Output file of ranges}
}
\description{
Convert SNPs to ranges
}
